// turborepo/shared/node-fabrik/v0.1.0/src/types.d.ts
export type NodePort = {
    index?: number
    plug?: 'input' | 'output'
    type: Type
    name: string
    displayName: string
    group: string
    default?: string | boolean | number
    tooltip?: string
    required?: boolean
    dependsOn?: readonly DependsOn[]
    isObject?: boolean
}

export type DependsOn = {
    name: string
    value: boolean | string
}

export type NoodlEnum = {
    label: string
    value: string
}

export type Type =
    | '*'
    | 'object'
    | 'array'
    | 'string'
    | 'stringlist'
    | 'number'
    | 'boolean'
    | 'signal'
    | 'enum'
    | 'color'
    | 'image'
    | 'icon'
    | 'font'
    | 'textStyle'
    | 'component'
    | 'dimension'
    | 'source'
    | 'resizing'
    | 'variable'
    | 'curve'
    | 'query-filter'
    | 'query-sorting'
    | 'pages'
    | 'proplist'
    | {
        name: Type;
        codeeditor?: 'javascript' | 'graphql' | 'html' | 'css';
        allowEditOnly?: boolean;
        allowConnectionsOnly?: boolean,
        enums?: NoodlEnum[],
        units?: string[],
        defaultUnit?: string
    }